Create README.md
Why would a review article on theoretical physics end up on GitHub, and not just on Arxiv? The idea is that review articles would be better if they were collaboratively written. An interesting experiment was done with the [review of AdS/CFT integrability](http://arxiv.org/abs/arXiv:1012.3982), where dozens of authors each contributed a relatively small piece. Here we test another approach: to open a single-authored text to collaboration by anyone interested. 

The BIB file contains hundreds of references, many more than are really used. The superfluous references have not been removed, because some of them could be useful later, in particular for versions that would include guides to the literature.